✨ feat(Message/room): room select command
Add a new command for selecting a discussion room.

If a message contains a number between 1 and 4, the bot will send a message with the available time slots and update the user's history to indicate that they are in the process of selecting a time.

If the message does not contain a valid number, the bot will send a message indicating that the room selection is invalid.