♻️ refactor: Use checkData function to get name
This commit refactors the code to use the checkData function to get the user's name.
This change is necessary because the previous code was not using the checkData function to get the user's name, which could lead to errors.
The checkData function is a more reliable way to get the user's name, and this change will help to improve the accuracy of the code.