🚑 fix: handle empty input for message menu
This commit fixes an issue where the message menu would be broken if the user entered an empty input.

The issue was caused by the fact that the code did not properly handle the case where the user entered an empty input. This resulted in the menu being broken and the user being unable to use it.

This commit fixes the issue by adding a check for an empty input. If the user enters an empty input, the menu will now be displayed with an error message.