✨ feat: Add functionality to upload documents to Turnitin
This commit introduces the ability to upload documents to Turnitin. When a user sends a supported document file, the bot will prompt the user to confirm the upload.

Once confirmed, the bot will upload the document to Turnitin and provide the user with a status update.

This functionality will allow users to easily submit their documents to Turnitin for plagiarism checking.